Created WorkoutManager, WorkoutCalendar
WorkoutManager is currently able to keep track of days in the workoutcalendar struct in the WorkoutCalendar file and store consecutive days with identical programming. This is in place of the WorkoutInfo/WorkoutData we had in our code before. The MonthView UI updates DayCell objects to a red color when this happens. Also refactored DayView to make the button objects created in a loop.